syntax: support multi-line syntax matches.
* lib/zile/syntax.c (state.init): Store the initial state for the
parser in the first line of the file when first called.
(state.new): Copy the stored parser state for this line into the
new state table.
(metatable.store): After highlighting, save the current state to
the next line in the buffer so that we can pick up highlighting
from there without re-parsing the entire file.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>