MutableString: fix infinite loop in rfind implementation.
Unfortunately, there remains an off-by-one issue in one or
more callers. For now, marked as "issue #3" and we continue
to accept the out-of-bounds argument.
* lib/zile/specs/MutableString_spec.yaml (rchr, rfind): Specify
behaviour of rchr and rfind methods.
* lib/zile/MutableString.lua (rchr, rfind): Don't engage the
search loop if the initial positions are out-of-bounds.
(rfind): Decrement index position when restarting find after
an unsuccessful match, to avoid an infinite loop!

Signed-off-by: Gary V. Vaughan <gary@gnu.org>